Done Revision
1. Speaker jadi Helper V
2. Pisahin function di lowercased subscriber view controller biar ga kepanjangan V
3. if else jadiin beda line V
4. Perhatiin weak self V
5. “View did load jangan panjang panjang” - Adhella, 2022 V
