+++ Gitra/Controller/Chord Detail/ChordDetailViewModel.swift
@@ -77,7 +77,9 @@ class ChordDetailViewModel{
                 isFinal = result.isFinal
                 
             }
-            else if result == nil { print("Result is nil") }
+            else if result == nil {
+                print("Result is nil")
+            }
             if isFinal {
                 print("Done recognizing speech")
             } else if error == nil {
@@ -217,36 +219,6 @@ class ChordDetailViewModel{
     
 }
 
-//MARK: - Speaker Class
 
-class Speaker: NSObject {
-    let synth = AVSpeechSynthesizer()
-    var note = ""
-    
-    override init() {
-        super.init()
-        synth.delegate = self
-    }
-    
-    func speak(_ str: String, playNote: String){
-        let utterance = AVSpeechUtterance(string: str)
-        utterance.voice = AVSpeechSynthesisVoice(language: "en-US")
-        utterance.rate = AVSpeechUtteranceMaximumSpeechRate / 2.0
-        synth.speak(utterance)
-        note = playNote
-    }
-    
-    func stop() {
-        synth.stopSpeaking(at: .immediate)
-    }
-}
-
-extension Speaker: AVSpeechSynthesizerDelegate {
-    func speechSynthesizer(_ synthesizer: AVSpeechSynthesizer, didFinish utterance: AVSpeechUtterance) {
-        if note != "" {
-            NotesMapping.shared.playSound(note)
-        }
-    }
-}
