+++ Gitra/Helper/Speaker.swift
@@ -1,0 +1,40 @@
+//
+//  Speaker.swift
+//  Gitra
+//
+//  Created by Samuel Maynard on 26/04/22.
+//
+import Foundation
+import AVFoundation
+
+//MARK: - Speaker Class
+
+class Speaker: NSObject {
+    let synth = AVSpeechSynthesizer()
+    var note = ""
+    
+    override init() {
+        super.init()
+        synth.delegate = self
+    }
+    
+    func speak(_ str: String, playNote: String){
+        let utterance = AVSpeechUtterance(string: str)
+        utterance.voice = AVSpeechSynthesisVoice(language: "en-US")
+        utterance.rate = AVSpeechUtteranceMaximumSpeechRate / 2.0
+        synth.speak(utterance)
+        note = playNote
+    }
+    
+    func stop() {
+        synth.stopSpeaking(at: .immediate)
+    }
+}
+
+extension Speaker: AVSpeechSynthesizerDelegate {
+    func speechSynthesizer(_ synthesizer: AVSpeechSynthesizer, didFinish utterance: AVSpeechUtterance) {
+        if note != "" {
+            NotesMapping.shared.playSound(note)
+        }
+    }
+}
